+++ common/src/main/java/io/github/flemmli97/flan/api/permission/ObjectToPermissionMap.java
@@ -8,6 +8,7 @@ import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.BoatItem;
 import net.minecraft.world.item.BoneMealItem;
 import net.minecraft.world.item.BucketItem;
@@ -42,6 +43,7 @@ import net.minecraft.world.level.block.TrapDoorBlock;
 import net.minecraft.world.level.block.TurtleEggBlock;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Consumer;
@@ -59,11 +61,14 @@ public class ObjectToPermissionMap {
     private static final Map<Item, ClaimPermission> itemToPermission = new HashMap<>();
     private static final Map<Predicate<Item>, Supplier<ClaimPermission>> itemPermissionBuilder = new HashMap<>();
 
+    private static final Map<EntityType<?>, ClaimPermission> entityToPermission = new HashMap<>();
+
     private static final Map<Block, ClaimPermission> leftClickBlockPermission = new HashMap<>();
 
     public static void reload(MinecraftServer server) {
         blockToPermission.clear();
         itemToPermission.clear();
+        entityToPermission.clear();
         leftClickBlockPermission.clear();
         for (Block block : CrossPlatformStuff.INSTANCE.registryBlocks().getIterator()) {
             blockPermissionBuilder.entrySet().stream().filter(e -> e.getKey().test(block)).map(Map.Entry::getValue).findFirst().ifPresent(sub -> blockToPermission.put(block, sub.get()));
@@ -71,58 +76,29 @@ public class ObjectToPermissionMap {
         for (Item item : CrossPlatformStuff.INSTANCE.registryItems().getIterator()) {
             itemPermissionBuilder.entrySet().stream().filter(e -> e.getKey().test(item)).map(Map.Entry::getValue).findFirst().ifPresent(sub -> itemToPermission.put(item, sub.get()));
         }
-        for (String s : ConfigHandler.config.itemPermission) {
+        process(ConfigHandler.config.itemPermission, BuiltInRegistries.ITEM, itemToPermission);
+        process(ConfigHandler.config.blockPermission, BuiltInRegistries.BLOCK, blockToPermission);
+        process(ConfigHandler.config.entityPermission, BuiltInRegistries.ENTITY_TYPE, entityToPermission);
+        process(ConfigHandler.config.leftClickBlockPermission, BuiltInRegistries.BLOCK, leftClickBlockPermission);
+    }
+
+    private static <T> void process(List<String> list, Registry<T> registry, Map<T, ClaimPermission> map) {
+        for (String s : list) {
             String[] sub = s.split("-");
             boolean remove = sub[1].equals("NONE");
             if (s.startsWith("@")) {
                 ResourceLocation res = new ResourceLocation(sub[0].substring(1));
-                processTag(res, BuiltInRegistries.ITEM, i -> {
+                processTag(res, registry, b -> {
                     if (remove)
-                        itemToPermission.remove(i);
+                        map.remove(b);
                     else
-                        itemToPermission.put(i, PermissionRegistry.get(sub[1]));
+                        map.put(b, PermissionRegistry.get(sub[1]));
                 });
             } else {
                 if (remove)
-                    itemToPermission.remove(CrossPlatformStuff.INSTANCE.registryItems().getFromId(new ResourceLocation(sub[0])));
+                    map.remove(registry.get(new ResourceLocation(sub[0])));
                 else
-                    itemToPermission.put(CrossPlatformStuff.INSTANCE.registryItems().getFromId(new ResourceLocation(sub[0])), PermissionRegistry.get(sub[1]));
-            }
-        }
-        for (String s : ConfigHandler.config.blockPermission) {
-            String[] sub = s.split("-");
-            boolean remove = sub[1].equals("NONE");
-            if (s.startsWith("@")) {
-                ResourceLocation res = new ResourceLocation(sub[0].substring(1));
-                processTag(res, BuiltInRegistries.BLOCK, b -> {
-                    if (remove)
-                        blockToPermission.remove(b);
-                    else
-                        blockToPermission.put(b, PermissionRegistry.get(sub[1]));
-                });
-            } else {
-                if (remove)
-                    blockToPermission.remove(CrossPlatformStuff.INSTANCE.registryBlocks().getFromId(new ResourceLocation(sub[0])));
-                else
-                    blockToPermission.put(CrossPlatformStuff.INSTANCE.registryBlocks().getFromId(new ResourceLocation(sub[0])), PermissionRegistry.get(sub[1]));
-            }
-        }
-        for (String s : ConfigHandler.config.leftClickBlockIgnore) {
-            String[] sub = s.split("-");
-            boolean remove = sub[1].equals("NONE");
-            if (s.startsWith("@")) {
-                ResourceLocation res = new ResourceLocation(sub[0].substring(1));
-                processTag(res, BuiltInRegistries.BLOCK, b -> {
-                    if (remove)
-                        leftClickBlockPermission.remove(b);
-                    else
-                        leftClickBlockPermission.put(b, PermissionRegistry.get(sub[1]));
-                });
-            } else {
-                if (remove)
-                    leftClickBlockPermission.remove(CrossPlatformStuff.INSTANCE.registryBlocks().getFromId(new ResourceLocation(sub[0])));
-                else
-                    leftClickBlockPermission.put(CrossPlatformStuff.INSTANCE.registryBlocks().getFromId(new ResourceLocation(sub[0])), PermissionRegistry.get(sub[1]));
+                    map.put(registry.get(new ResourceLocation(sub[0])), PermissionRegistry.get(sub[1]));
             }
         }
     }
@@ -141,6 +117,10 @@ public class ObjectToPermissionMap {
 
     public static ClaimPermission getFromItem(Item item) {
         return itemToPermission.get(item);
+    }
+
+    public static ClaimPermission getFromEntity(EntityType<?> entity) {
+        return entityToPermission.get(entity);
     }
 
     public static ClaimPermission getForLeftClickBlock(Block block) {